Call our linting logic on document change
